backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
openai>=1.25.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
pyngrok>=7.1.0
